Move constants to mf directory
