+++ src/mars_patcher/locations.py
@@ -6,7 +6,8 @@ from typing import TYPE_CHECKING, ClassVar
 
 from frozendict import frozendict
 
-from mars_patcher.constants.items import (
+from mars_patcher.data import get_data_path
+from mars_patcher.mf.constants.items import (
     ITEM_ENUMS,
     ITEM_SPRITE_ENUMS,
     JINGLE_ENUMS,
@@ -34,7 +35,6 @@ from mars_patcher.constants.items import (
     ItemType,
     MajorSource,
 )
-from mars_patcher.data import get_data_path
 from mars_patcher.text import Language
 
 if TYPE_CHECKING:
