+++ src/mars_patcher/mf/constants/minimap_tiles.py
