+++ src/mars_patcher/mf/constants/palettes.py
