+++ src/mars_patcher/misc_patches.py
@@ -1,6 +1,6 @@
-import mars_patcher.constants.game_data as gd
-from mars_patcher.constants.reserved_space import ReservedConstants
+import mars_patcher.mf.constants.game_data as gd
 from mars_patcher.data import get_data_path
+from mars_patcher.mf.constants.reserved_space import ReservedConstants
 from mars_patcher.patching import BpsDecoder, IpsDecoder
 from mars_patcher.rom import Rom
