+++ src/mars_patcher/titlescreen_text.py
@@ -1,5 +1,5 @@
-from mars_patcher.constants.reserved_space import ReservedConstants
 from mars_patcher.mf.auto_generated_types import MarsschemamfTitletextItem
+from mars_patcher.mf.constants.reserved_space import ReservedConstants
 from mars_patcher.rom import Rom
 
 TITLE_TEXT_POINTER_ADDR = ReservedConstants.TITLESCREEN_TEXT_POINTERS_POINTER_ADDR
